+++ tests/serpapi_test.ts
@@ -28,6 +28,7 @@ import {
   getLocations,
   getMd,
   getMdBySearchId,
+  HTTPError,
   InvalidArgumentError,
   InvalidTimeoutError,
   MissingApiKeyError,
@@ -248,12 +249,15 @@ describe("uploadImage", () => {
     const executeStub = stub(
       _internals,
       "uploadImage",
-      () => Promise.reject(apiError),
+      () => Promise.reject(new HTTPError(401, apiError)),
     );
     config.api_key = "test_api_key";
     try {
       const error = await uploadImage({ image }).catch((error) => error);
-      assertEquals(error, apiError);
+      assertInstanceOf(error, HTTPError);
+      assertEquals(error.statusCode, 401);
+      assertEquals(error.body, apiError);
+      assertEquals(error.message, "Invalid image");
     } finally {
       executeStub.restore();
     }